+++ frontend/src/features/mapSideBar/tabs/propertyDetails/update/UpdatePropertyDetailsForm.tsx
@@ -2,6 +2,7 @@ import { Button } from 'components/common/buttons';
 import { Input, Multiselect, Select, Text, TextArea } from 'components/common/form';
 import { RadioGroup } from 'components/common/form/RadioGroup';
 import { YesNoSelect } from 'components/common/form/YesNoSelect';
+import { Scrollable } from 'components/common/Scrollable/Scrollable';
 import * as API from 'constants/API';
 import { PropertyAdjacentLandTypes, PropertyTenureTypes } from 'constants/index';
 import { Form, FormikProps, getIn } from 'formik';
@@ -9,7 +10,6 @@ import { useLookupCodeHelpers } from 'hooks/useLookupCodeHelpers';
 import Api_TypeCode from 'models/api/TypeCode';
 import React from 'react';
 import { ButtonToolbar, Col, Row } from 'react-bootstrap';
-import { useHistory } from 'react-router-dom';
 import styled from 'styled-components';
 import { stringToBoolean } from 'utils/formUtils';
 
@@ -20,10 +20,15 @@ import { LandMeasurementTable } from './components/LandMeasurementTable';
 import { VolumetricMeasurementTable } from './components/VolumetricMeasurementTable';
 import { UpdatePropertyDetailsFormModel } from './models';
 
-export const UpdatePropertyDetailsForm: React.FC<FormikProps<UpdatePropertyDetailsFormModel>> = ({
+export interface IUpdatePropertyDetailsFormProps
+  extends FormikProps<UpdatePropertyDetailsFormModel> {
+  onCancel: () => void;
+}
+
+export const UpdatePropertyDetailsForm: React.FC<IUpdatePropertyDetailsFormProps> = ({
   values,
+  onCancel,
 }) => {
-  const location = useHistory();
   // Lookup codes
   const { getByType, getOptionsByType } = useLookupCodeHelpers();
   const anomalyOptions = getByType(API.PROPERTY_ANOMALY_TYPES);
@@ -48,148 +53,146 @@ export const UpdatePropertyDetailsForm: React.FC<FormikProps<UpdatePropertyDetai
   const volumetricMeasurement = getIn(values, 'volumetricMeasurement') as number;
   const volumetricUnit = getIn(values, 'volumetricUnitTypeCode') as string;
 
-  const onCancel = () => {
-    location.goBack();
-  };
-
   return (
     <StyledForm>
-      <Section header="Property attributes">
-        <SectionField label="MOTI region">
-          <Text field="motiRegion.REGION_NAME" />
-        </SectionField>
-        <SectionField label="Highways district">
-          <InlineContainer>
-            <Text field="highwaysDistrict.DISTRICT_NUMBER" />
-            {'-'}
-            <Text field="highwaysDistrict.DISTRICT_NAME" />
-          </InlineContainer>
-        </SectionField>
-        <SectionField label="Electoral district">
-          <Text field="electoralDistrict.ED_NAME" />
-        </SectionField>
-        <SectionField label="Agricultural Land Reserve">
-          <Text>{values.isALR ? 'Yes' : 'No'}</Text>
-        </SectionField>
-        <SectionField label="Land parcel type">{/* TODO */}</SectionField>
-        <SectionField label="Municipal zoning">
-          <Input field="municipalZoning" />
-        </SectionField>
-        <SectionField label="Anomalies">
-          <Multiselect
-            field="anomalies"
-            displayValue="name"
-            placeholder=""
-            hidePlaceholder
-            options={anomalyOptions}
-          />
-        </SectionField>
-      </Section>
-
-      <Section header="Tenure Status">
-        <SectionField label="Tenure status">
-          <Multiselect
-            field="tenures"
-            displayValue="name"
-            placeholder=""
-            hidePlaceholder
-            options={tenureOptions}
-          />
-        </SectionField>
-        <SectionField label="Provincial Public Hwy">
-          <YesNoSelect field="isProvincialPublicHwy"></YesNoSelect>
-        </SectionField>
-        {isHighwayRoad && (
-          <SectionField label="Highway / Road">
+      <Content vertical>
+        <Section header="Property attributes">
+          <SectionField label="MOTI region">
+            <Text field="motiRegion.REGION_NAME" />
+          </SectionField>
+          <SectionField label="Highways district">
+            <InlineContainer>
+              <Text field="highwaysDistrict.DISTRICT_NUMBER" />
+              {'-'}
+              <Text field="highwaysDistrict.DISTRICT_NAME" />
+            </InlineContainer>
+          </SectionField>
+          <SectionField label="Electoral district">
+            <Text field="electoralDistrict.ED_NAME" />
+          </SectionField>
+          <SectionField label="Agricultural Land Reserve">
+            <Text>{values.isALR ? 'Yes' : 'No'}</Text>
+          </SectionField>
+          <SectionField label="Land parcel type">{/* TODO */}</SectionField>
+          <SectionField label="Municipal zoning">
+            <Input field="municipalZoning" />
+          </SectionField>
+          <SectionField label="Anomalies">
             <Multiselect
-              field="roadTypes"
+              field="anomalies"
               displayValue="name"
               placeholder=""
               hidePlaceholder
-              options={roadTypeOptions}
+              options={anomalyOptions}
             />
-          </SectionField>
-        )}
-        {isAdjacentLand && (
-          <SectionField label="Adjacent land">
-            <Multiselect
-              field="adjacentLands"
-              displayValue="name"
-              placeholder=""
-              hidePlaceholder
-              options={adjacentLandOptions}
-            />
-          </SectionField>
-        )}
-      </Section>
-
-      {isIndianReserve && (
-        <Section header="First Nations Information">
-          <SectionField label="Band name">
-            <Text field="firstNations.bandName" />
-          </SectionField>
-          <SectionField label="Reserve name">
-            <Text field="firstNations.reserveName" />
           </SectionField>
         </Section>
-      )}
 
-      <Section header="Area">
-        <Row>
-          <Col>
-            <Row>
-              <Col className="col-10">
-                <LandMeasurementTable area={landArea} areaUnitTypeCode={areaUnit} />
-              </Col>
-            </Row>
-          </Col>
-          <LeftBorderCol>
-            <StyledFieldLabel>Is this a volumetric parcel?</StyledFieldLabel>
-            <RadioGroup
-              flexDirection="row"
-              field="isVolumetricParcel"
-              radioGroupClassName="pb-3"
-              radioValues={[
-                {
-                  radioLabel: 'Yes',
-                  radioValue: 'true',
-                },
-                {
-                  radioLabel: 'No',
-                  radioValue: 'false',
-                },
-              ]}
+        <Section header="Tenure Status">
+          <SectionField label="Tenure status">
+            <Multiselect
+              field="tenures"
+              displayValue="name"
+              placeholder=""
+              hidePlaceholder
+              options={tenureOptions}
             />
+          </SectionField>
+          <SectionField label="Provincial Public Hwy">
+            <YesNoSelect field="isProvincialPublicHwy"></YesNoSelect>
+          </SectionField>
+          {isHighwayRoad && (
+            <SectionField label="Highway / Road">
+              <Multiselect
+                field="roadTypes"
+                displayValue="name"
+                placeholder=""
+                hidePlaceholder
+                options={roadTypeOptions}
+              />
+            </SectionField>
+          )}
+          {isAdjacentLand && (
+            <SectionField label="Adjacent land">
+              <Multiselect
+                field="adjacentLands"
+                displayValue="name"
+                placeholder=""
+                hidePlaceholder
+                options={adjacentLandOptions}
+              />
+            </SectionField>
+          )}
+        </Section>
 
-            {isVolumetricParcel && (
-              <>
-                <SectionField label="Type">
-                  <Select
-                    field="volumetricParcelTypeCode"
-                    options={volumetricTypeOptions}
-                    placeholder={values.volumetricParcelTypeCode ? undefined : 'Please Select'}
-                  />
-                </SectionField>
+        {isIndianReserve && (
+          <Section header="First Nations Information">
+            <SectionField label="Band name">
+              <Text field="firstNations.bandName" />
+            </SectionField>
+            <SectionField label="Reserve name">
+              <Text field="firstNations.reserveName" />
+            </SectionField>
+          </Section>
+        )}
 
-                <Row>
-                  <Col className="col-10">
-                    <VolumetricMeasurementTable
-                      volume={volumetricMeasurement}
-                      volumeUnitTypeCode={volumetricUnit}
+        <Section header="Area">
+          <Row>
+            <Col>
+              <Row>
+                <Col className="col-10">
+                  <LandMeasurementTable area={landArea} areaUnitTypeCode={areaUnit} />
+                </Col>
+              </Row>
+            </Col>
+            <LeftBorderCol>
+              <StyledFieldLabel>Is this a volumetric parcel?</StyledFieldLabel>
+              <RadioGroup
+                flexDirection="row"
+                field="isVolumetricParcel"
+                radioGroupClassName="pb-3"
+                radioValues={[
+                  {
+                    radioLabel: 'Yes',
+                    radioValue: 'true',
+                  },
+                  {
+                    radioLabel: 'No',
+                    radioValue: 'false',
+                  },
+                ]}
+              />
+
+              {isVolumetricParcel && (
+                <>
+                  <SectionField label="Type">
+                    <Select
+                      field="volumetricParcelTypeCode"
+                      options={volumetricTypeOptions}
+                      placeholder={values.volumetricParcelTypeCode ? undefined : 'Please Select'}
                     />
-                  </Col>
-                </Row>
-              </>
-            )}
-          </LeftBorderCol>
-        </Row>
-      </Section>
+                  </SectionField>
 
-      <Section header="Notes">
-        <TextArea field="notes" rows={4} />
-      </Section>
+                  <Row>
+                    <Col className="col-10">
+                      <VolumetricMeasurementTable
+                        volume={volumetricMeasurement}
+                        volumeUnitTypeCode={volumetricUnit}
+                      />
+                    </Col>
+                  </Row>
+                </>
+              )}
+            </LeftBorderCol>
+          </Row>
+        </Section>
 
-      <Row className="m-0 justify-content-md-end">
+        <Section header="Notes">
+          <TextArea field="notes" rows={4} />
+        </Section>
+      </Content>
+
+      <Footer>
         <ButtonToolbar className="cancelSave">
           <Button className="mr-5" variant="secondary" type="button" onClick={onCancel}>
             Cancel
@@ -198,9 +201,24 @@ export const UpdatePropertyDetailsForm: React.FC<FormikProps<UpdatePropertyDetai
             Save
           </Button>
         </ButtonToolbar>
-      </Row>
+      </Footer>
     </StyledForm>
   );
 };
 
 const StyledForm = styled(Form)``;
+
+const Content = styled(Scrollable)`
+  display: flex;
+  flex-direction: column;
+  flex-grow: 1;
+  text-align: left;
+  height: 100%;
+  padding-right: 1rem;
+  padding-bottom: 1rem;
+`;
+
+const Footer = styled(Row)`
+  margin: 0;
+  justify-content: end;
+`;
